+++ pom.xml
@@ -61,7 +61,7 @@
                     </execution>
                 </executions>
             </plugin>
-            
+
         </plugins>
     </build>
