+++ src/main/java/me/pr0crustes/frontend/gui/scene/CropController.java
@@ -7,10 +7,10 @@ import javafx.scene.control.TextField;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
-import me.pr0crustes.backend.classes.FileSelector;
-import me.pr0crustes.backend.classes.Numbers;
-import me.pr0crustes.backend.classes.PDFCropper;
-import me.pr0crustes.backend.classes.PDFManager;
+import me.pr0crustes.backend.classes.file.FileSelector;
+import me.pr0crustes.backend.classes.number.Numbers;
+import me.pr0crustes.backend.classes.pdf.PDFCropper;
+import me.pr0crustes.backend.classes.pdf.PDFManager;
 import me.pr0crustes.backend.enums.FileExtensions;
 import me.pr0crustes.backend.exeptions.ArgumentException;
 import me.pr0crustes.backend.exeptions.NoFileException;
