+++ src/main/java/me/pr0crustes/frontend/gui/scene/InsertController.java
@@ -9,10 +9,10 @@ import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Font;
-import me.pr0crustes.backend.classes.FileSelector;
-import me.pr0crustes.backend.classes.Numbers;
-import me.pr0crustes.backend.classes.PDFInsert;
-import me.pr0crustes.backend.classes.PDFManager;
+import me.pr0crustes.backend.classes.file.FileSelector;
+import me.pr0crustes.backend.classes.number.Numbers;
+import me.pr0crustes.backend.classes.pdf.PDFInsert;
+import me.pr0crustes.backend.classes.pdf.PDFManager;
 import me.pr0crustes.backend.enums.FileExtensions;
 import me.pr0crustes.backend.exeptions.ArgumentException;
 import me.pr0crustes.backend.exeptions.NoFileException;
